style moved runner to own folder
